PO-1564 - DefendantAccountAtAGlance - DTO display names
Follows JSON schema update to include the display name
- InstalmentPeriod
- PaymentTermsType
- LanguagePreference

+++ src/main/java/uk/gov/hmcts/opal/dto/common/Enforcer.java
@@ -14,7 +14,7 @@ import uk.gov.hmcts.opal.dto.ToJsonString;
 public class Enforcer implements ToJsonString {
 
     @JsonProperty("enforcer_id")
-    private Integer enforcerId;
+    private Long enforcerId;
 
     @JsonProperty("enforcer_name")
     private String enforcerName;
